requirements/training.txt
torch
torchvision
timm
PyYAML
openslide-python
Pillow
huggingface_hub
safetensors
